Disable launchMain() on SelectHomeSlide
Disable only if the current slide is SelectHomeSlide and the House of YACS
setting has not been set yet.

This is to prevent the user from being able to skip to the main activity
without first setting the YACS API base url.

This is required in order to ensure YACS Android consistency.